+++ R/predict.R
@@ -33,8 +33,8 @@ predict.tinyVAST <-
 function( object,
           newdata,
           remove_origdata = FALSE,
-          what = c("mu_g", "p_g", "palpha_g", "pgamma_g", "pepsilon_g", "pomega_g", "pxi_g",
-                   "p2_g", "palpha2_g", "pgamma2_g", "pepsilon2_g", "pomega2_g", "pxi2_g"),
+          what = c("mu_g", "p_g", "palpha_g", "pgamma_g", "pepsilon_g", "pomega_g", "pdelta_g", "pxi_g",
+                   "p2_g", "palpha2_g", "pgamma2_g", "pepsilon2_g", "pomega2_g", "pdelta2_g", "pxi2_g"),
           se.fit = FALSE,
           ... ){
 
@@ -524,7 +524,7 @@ function( object,
   }
 
   # Check for obvious issues ... no NAs except in RAMstart
-  index_drop = match(c("ram_space_term_start","ram_spacetime_term_start","ram2_space_term_start","ram2_spacetime_term_start"),names(tmb_data2))
+  index_drop = match(c("ram_space_term_start","ram_time_term_start","ram_spacetime_term_start","ram2_space_term_start","ram2_time_term_start","ram2_spacetime_term_start"),names(tmb_data2))
   if( any(is.na(tmb_data2[-index_drop])) ){
     stop("Check output of `add_predictions` for NAs")
   }